% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{fileOneDrive}
\alias{fileOneDrive}
\title{Microsoft OneDrive File Upload Control}
\usage{
fileOneDrive(inputId, app, multiple = F, accept = NULL,
  viewType = "files", queryParameters = NULL,
  buttonLabel = "Browse...", icon = shiny::icon("cloud"),
  width = NULL, endpointHint = "api.onedrive.com", sdk = 7.2, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{multiple}{Whether the user should be allowed to select and upload
multiple files at once. \bold{Does not work on older browsers, including
Internet Explorer 9 and earlier.}}

\item{accept}{A character vector of MIME types; gives the browser a hint of
what kind of files the server is expecting.}

\item{viewType}{The type of item that can be selected. Can be \code{"files"}, \code{"folders"} or \code{"all"}}

\item{queryParameters}{A character vector of additional query parameters as specified by the OneDrive API that define how an item is returned.
This typically includes a select and/or expand value}

\item{buttonLabel}{The label used on the button. Can be text or an HTML tag
object.}

\item{icon}{An optional \code{\link{icon}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}

\item{endpointHint}{Endpoint hint is used for SDK redirects the app to the right OAuth endpoint based on which OneDrive API endpoints the app wants talk to.
The value of \code{endpointHint} could be \code{"api.onedrive.com"} for OneDrive personal, the OneDrive for Business URL or a SharePoint document library URL,
ex. \code{"https://contoso-my.sharepoint.com/personal/foo_contoso_onmicrosoft_com/"} or \code{"https://contoso.sharepoint.com/shared\%20documents/"}}

\item{sdk}{SDK version}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
A Shiny button. Files selected will get returned as json to input$inputId
}
\description{
Shiny button that when clicked shows a file picker used to get Graph file IDs and metadata.
See \href{https://docs.microsoft.com/en-us/onedrive/developer/controls/file-pickers/js-v72/open-file}{File Picker reference guide} for details
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  br(),
  fileOneDrive("selected",rainy_app("[client ID]")),
  br(),
  verbatimTextOutput("response")
)

server <- function(input,output) {
  output$response <- renderPrint(input$selected)
}

runApp(shinyApp(ui=ui,server=server),launch.browser=T)
}
}
