% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_endpoint}
\alias{graph_endpoint}
\title{Microsoft Graph Endpoint}
\usage{
graph_endpoint(tenant = "common", ...)
}
\arguments{
\item{tenant}{Graph tenant. For single-tenant should be something like \code{"contoso.onmicrosoft.com"}, otherwise use default \code{"common"}}

\item{...}{Not used. Arguments passed from other functions}
}
\description{
Generates an endpoint object that can be used in \code{\link{graph_login}}
}
