% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_get}
\alias{graph_get}
\title{Microsoft Graph GET Request}
\usage{
graph_get(endpoint, ...)
}
\arguments{
\item{endpoint}{URL endpoint. See \href{https://developer.microsoft.com/en-us/graph/docs/concepts/overview}{API overview} for examples}

\item{...}{Optional arguments passed to \code{\link{graph_url}} and \code{\link{graph_handler}}}
}
\value{
List of vectors and a data.frame of values
}
\description{
Standard GET call to Graph API
}
\examples{
\dontrun{
graph_get("me")
graph_get("drives")
}
}
