% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_handler}
\alias{graph_handler}
\title{Microsoft Graph API Handler}
\usage{
graph_handler(url, ..., fileext = "", sheet = 1, skip = 0,
  col_types = NULL, verb = "GET")
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{...}{Optional arguments passed to \code{\link[httr]{VERB}}}

\item{fileext}{a non-empty character vector giving the file extension}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{skip}{Minimum number of rows to skip before reading anything, be it
column names or data. Leading empty rows are automatically skipped, so this
is a lower bound. Ignored if \code{range} is given.}

\item{col_types}{Either \code{NULL} to guess all from the spreadsheet or a
character vector containing one entry per column from these options:
"skip", "guess", "logical", "numeric", "date", "text" or "list". If exactly
one \code{col_type} is specified, it will be recycled. The content of a cell in
a skipped column is never read and that column will not appear in the data
frame output. A list cell loads a column as a list of length 1 vectors,
which are typed using the type guessing logic from \code{col_types = NULL}, but
on a cell-by-cell basis.}

\item{verb}{Name of verb to use.}
}
\description{
Wrapper for handling API calls/responses
}
