% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_login}
\alias{graph_login}
\title{Microsoft Graph Login}
\usage{
graph_login(app, code = NULL, scope = "Files.ReadWrite.All",
  endpoint = graph_endpoint())
}
\arguments{
\item{app}{An OAuth consumer application, created by
\code{\link[=oauth_app]{oauth_app()}}}

\item{scope}{a character vector of scopes to request.}

\item{endpoint}{An OAuth endpoint, created by \code{\link[=oauth_endpoint]{oauth_endpoint()}}}
}
\description{
Performs OAuth2.0 and stores a token for later calls to Graph API
}
