% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_search}
\alias{graph_search}
\title{Microsoft Graph Search Query}
\usage{
graph_search(text, drive = "me/drive/root", top = NA_integer_, ...)
}
\arguments{
\item{text}{The query text used to search for items. Values may be matched across several fields including filename, metadata, and file content}

\item{drive}{Drive item endpoint preceding "/search".
See \href{https://docs.microsoft.com/en-us/onedrive/developer/rest-api/api/driveitem_get}{Get file or folder} for examples}

\item{top}{Number of items to return in a result set. If \code{NA} will return all}

\item{...}{Optional arguments passed to \code{\link{graph_get}}}
}
\value{
data.table of query results
}
\description{
Search the hierarchy of drive items for items matching a query. Looks at file metadata and data within files
}
