% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_upload}
\alias{graph_upload}
\title{Microsoft Graph File Upload}
\usage{
graph_upload(path, endpoint = "me/drive/root",
  name = curl::curl_escape(basename(path)),
  fileext = tools::file_ext(path), ...)
}
\arguments{
\item{endpoint}{URL endpoint. See \href{https://developer.microsoft.com/en-us/graph/docs/concepts/overview}{API overview} for examples}

\item{name}{Filename to save as. Defaults to original filename}

\item{fileext}{a non-empty character vector giving the file extension}
}
\value{
List of drive item IDs and metadata
}
\description{
Upload the contents of a new file or update the contents of an existing file
}
