% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_url}
\alias{graph_url}
\title{Microsoft Graph Base URL}
\usage{
graph_url(endpoint, api = 1, ...)
}
\arguments{
\item{endpoint}{URL endpoint. See \href{https://developer.microsoft.com/en-us/graph/docs/concepts/overview}{API overview} for examples}

\item{api}{API version. If \code{0} will use beta API}
}
\description{
Helper function to build a base URL with version number
}
