% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainy-package.R
\docType{package}
\name{rainy-package}
\alias{rainy}
\alias{rainy-package}
\title{rainy: Research Compendium for the Paper 'The relationship between rain intensity and Ascochyta blight in chickpea development'}
\description{
Documents the workflow and processes involved in the analysis of
   data for the paper '', <https://zenodo.org/badge/latestdoi/378052232>.
}
\author{
\strong{Maintainer}: Ihsanul Khaliq \email{ihsan.khaliq@usq.edu.au} (\href{https://orcid.org/0000-0003-4171-0917}{ORCID})

Authors:
\itemize{
  \item Kevin Moore \email{kevin.moore@dpi.nsw.gov.au} (\href{https://orcid.org/0000-0001-9643-6885}{ORCID})
  \item Adam H. Sparks \email{adam.sparks@usq.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID}) [conceptor]
}

Other contributors:
\itemize{
  \item Grains Research and Development Corporation (Provided support through funding GRDC Project
                     USQ1903-003RTX) [funder, copyright holder]
  \item Western Australia Agriculture Authority (Provided support through Adam Sparks' time.) [funder, copyright holder]
  \item University of Southern Queensland [copyright holder]
}

}
\keyword{internal}
