% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainy.R
\docType{package}
\name{rainy}
\alias{rainy}
\alias{rainy-package}
\title{Cloud API Wrappers and Shiny Tools}
\description{
API wrappers and tools to facilitate interations between cloud based databases and Shiny apps
}
