% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainy.R
\name{rainy_app}
\alias{rainy_app}
\title{OAuth App}
\usage{
rainy_app(key, redirect_uri = httr::oauth_callback())
}
\arguments{
\item{key}{consumer key, also sometimes called the client ID}

\item{redirect_uri}{The URL that user will be redirected to after
authorisation is complete. You should generally leave this as the default
unless you're using a non-standard auth flow (like with shiny).}
}
\description{
Helper function that generates an OAuth app object with different defaults
}
\details{
For Microsoft Graph, need to \href{https://apps.dev.microsoft.com/}{register an app}
with Azure to set permissions
}
