% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainy.R
\name{rainy_ui}
\alias{rainy_ui}
\title{Shiny OAuth Redirect}
\usage{
rainy_ui(ui, app, endpoint, scope = NULL)
}
\arguments{
\item{ui}{Shiny UI opbject}

\item{app}{An OAuth consumer application, created by
\code{\link[=oauth_app]{oauth_app()}}}

\item{endpoint}{An OAuth endpoint, created by \code{\link[=oauth_endpoint]{oauth_endpoint()}}}

\item{scope}{a character vector of scopes to request.}
}
\description{
Check for oauth credentials and store code in the URL. This allows for a Shiny to perform all susequent interactions with a cloud database instead
of having to login with each interaction (see \code{\link{fileOneDrive}} example)
}
\details{
Adapted from \href{https://gist.github.com/hadley/144c406871768d0cbe66b0b810160528}{Hadley's solution}
}
\examples{
\dontrun{
library(shiny)

app <- rainy_app("[client ID]") #remember to set redirect URI in Azure

ui <- rainy_ui(fluidPage(
  br(),
  h3("You are now logged in as"),
  br(),
  verbatimTextOutput("me")
),app,graph_endpoint(),"Files.ReadWrite.All")

server <- function(input,output,session) {
  params <- parseQueryString(isolate(session$clientData$url_search))
  if (is.null(params$code))
     return()
  graph_login(app,params$code)
  output$me <- renderPrint(graph_get("me"))
}

runApp(shinyApp(ui=ui,server=server),launch.browser=T)
}
}
