% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainy-package.R
\docType{package}
\name{rainy-package}
\alias{rainy}
\alias{rainy-package}
\title{rainy: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ihsanul Khaliq \email{ihsan.khaliq@usq.edu.au} (\href{https://orcid.org/0000-0003-4171-0917}{ORCID})

Authors:
\itemize{
  \item Adam H. Sparks \email{adam.sparks@usq.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
}

Other contributors:
\itemize{
  \item GRDC Project USQ1903-003RTX [funder]
}

}
\keyword{internal}
