% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainy-package.R
\docType{package}
\name{rainy-package}
\alias{rainy}
\alias{rainy-package}
\title{rainy: Research Compendium for the Paper ''}
\description{
Documents the workflow and processes involved in the analysis of
   data for the paper '', <Paper DOI goes here>.
}
\author{
\strong{Maintainer}: Ihsanul Khaliq \email{ihsan.khaliq@usq.edu.au} (\href{https://orcid.org/0000-0003-4171-0917}{ORCID})

Authors:
\itemize{
  \item Adam H. Sparks \email{adam.sparks@usq.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
}

Other contributors:
\itemize{
  \item GRDC Project USQ1903-003RTX [funder]
}

}
\keyword{internal}
